# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.CartResponse:
    properties:
      count:
        type: integer
      trip_id:
        type: integer
    type: object
...
  handler.ScenarioResponse:
    properties:
      aero_coeff:
        type: number
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      rolling_coeff:
        type: number
      speed:
        type: number
      status:
        type: string
      system_consumption:
        type: number
      type:
        type: string
    type: object
  handler.TripResponse:
    properties:
      completed_at:
        type: string
      created_at:
        description: Изменено на RussianTime
        type: string
      creator_login:
        type: string
      id:
        type: integer
      moderator_login:
        type: string
      remaining_charge:
        description: убрал omitempty
        type: number
      scenarios:
        items:
          $ref: '#/definitions/handler.TripScenarioResponse'
        type: array
      start_charge:
        type: number
      status:
        type: string
      submitted_at:
        type: string
    type: object
  handler.TripScenarioResponse:
    properties:
      duration:
        type: number
      scenario:
        $ref: '#/definitions/handler.ScenarioResponse'
      scenario_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@tesla-app.com
    name: API Support
    url: http://localhost:8080
  description: API for Tesla driving scenarios and trip management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Tesla App API
  version: "1.0"
paths:
...
  /api/trips:
    get:
      consumes:
      - application/json
      description: Get trips with filtering. For moderators - all trips, for users
        - only their trips
      parameters:
      - description: Status filter
        in: query
        name: status
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TripResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get trips list
      tags:
      - Trips
    post:
      consumes:
      - application/json
      description: Create new trip draft
      parameters:
      - description: Trip data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Trip created
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create trip
      tags:
      - Trips
  /api/trips/{trip_id}:
    delete:
      consumes:
      - application/json
      description: Delete trip (only by owner)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trip deleted
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Access denied
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete trip
      tags:
      - Trips
    get:
      consumes:
      - application/json
      description: Get trip details by ID. Users can only access their own trips,
        moderators can access all.
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TripResponse'
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Access denied
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get trip by ID
      tags:
      - Trips
    put:
      consumes:
      - application/json
      description: Update trip data (only for draft trips by owner)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Start charge data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Trip updated
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Access denied
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update trip
      tags:
      - Trips
  /api/trips/{trip_id}/reviewtrip:
    put:
      consumes:
      - application/json
      description: Complete or reject trip (moderator only)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Review action
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Review successful
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Review trip
      tags:
      - Trips
...
  /api/trips/{trip_id}/submittrip:
    put:
      consumes:
      - application/json
      description: Submit draft trip for moderation and start async charge calculation
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trip submitted
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "403":
          description: Cannot submit trip
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Submit trip
      tags:
      - Trips
  /api/trips/{trip_id}/update_charge:
    put:
      consumes:
      - application/json
      - application/json
      description: |-
        Update remaining charge for a trip (called by async calculator)
        Update remaining charge for a trip (called by async calculator)
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Charge data
        in: body
        name: input
        required: true
        schema:
          type: object
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Charge data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Charge updated
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Trip not found
          schema:
            type: object
      summary: Update trip charge
      tags:
      - Trips
      - Trips
  /api/trips/scenarioscart:
    get:
      consumes:
      - application/json
      description: Get user's scenarios cart count. For guests returns 0.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CartResponse'
      security:
      - BearerAuth: []
      summary: Get scenarios cart
      tags:
      - Trips
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"